Rename `gtest` to `gear-test`, and lib to `gtest`
