+++ .github/workflows/CI.yaml
@@ -140,7 +140,7 @@ jobs:
           npm run build
           cd ..
           RUST_LOG=gwasm=debug,gear_core_processor=debug ./target/release/gear-node --dev --ws-max-out-buffer-capacity 500 --rpc-max-payload 500 -l 0 & sleep 10
-          node rpc-tests/dist/index.js ./gtest/spec/*.yaml
+          node rpc-tests/dist/index.js ./gear-test/spec/*.yaml
 
       - name: "Cache: Pack"
         run: |
@@ -160,10 +160,10 @@ jobs:
           cp target/release/wbuild/gear-runtime/gear_runtime.compact.wasm artifact/
           cp target/release/wbuild/gear-runtime/gear_runtime.wasm artifact/
           cp target/release/gear-node artifact/
-          cp target/release/gtest artifact/
+          cp target/release/gear-test artifact/
           cp target/release/wasm-proc artifact/
           strip artifact/gear-node || true
-          strip artifact/gtest || true
+          strip artifact/gear-test || true
           strip artifact/wasm-proc || true
 
       - name: Upload artifacts
@@ -236,10 +236,10 @@ jobs:
 
       - name: "Cache: Unpack"
         run: |
-          cp /root/cache/gtest* /tmp/
-          tar -xf /tmp/gtest_cache.tar -C /
-          tar -xf /tmp/gtest_cargo_registry.tar -C /
-          tar -xf /tmp/gtest_target.tar
+          cp /root/cache/gear-test* /tmp/
+          tar -xf /tmp/gear-test_cache.tar -C /
+          tar -xf /tmp/gear-test_cargo_registry.tar -C /
+          tar -xf /tmp/gear-test_target.tar
 
       - name: "Install: Node.js packages"
         run: ./scripts/gear.sh init js
@@ -258,7 +258,7 @@ jobs:
 
       - name: "Cache: Pack"
         run: |
-          tar -cf /tmp/gtest_target.tar ./target
-          tar -cf /tmp/gtest_cache.tar /tmp/cachepot
-          tar -cf /tmp/gtest_cargo_registry.tar /tmp/cargo/registry/cache /tmp/cargo/registry/index /tmp/cargo/git
+          tar -cf /tmp/gear-test_target.tar ./target
+          tar -cf /tmp/gear-test_cache.tar /tmp/cachepot
+          tar -cf /tmp/gear-test_cargo_registry.tar /tmp/cargo/registry/cache /tmp/cargo/registry/index /tmp/cargo/git
           mv /tmp/*.tar /root/cache/
