--- gear-test-lib/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "gear-test-lib"
-version = "0.1.0"
-authors = ["Gear Technologies"]
-edition = "2018"
-license = "GPL-3.0"
-
-[dependencies]
-gear-core = { path = "../core" }
-gear-backend-wasmtime = { path = "../core-backend/wasmtime" }
-core-processor = { package = "gear-core-processor", path = "../core-processor" }
-
-codec = { package = "parity-scale-codec", version = "2", features = ["derive"] }
-hex = "0.4.3"
-colored = "2.0.0"
-env_logger = "0.9.0"
-path-clean = "0.1.0"
