--- gear-test-lib/src/lib.rs
@@ -1,6 +1,0 @@
-mod manager;
-mod program;
-mod system;
-
-pub use program::{Program, WasmProgram};
-pub use system::System;
