+++ gear-test/src/js/decode.js
