+++ gear-test/src/js/encode.js
