+++ gear-test/src/js/package.json
