+++ gtest/Cargo.toml
@@ -1,39 +1,17 @@
 [package]
-name = "gear-test"
+name = "gear-test-lib"
 version = "0.1.0"
 authors = ["Gear Technologies"]
 edition = "2018"
 license = "GPL-3.0"
 
 [dependencies]
-anyhow = "1"
-clap = { version = "3.0.10", features = ["derive"] }
-derive_more = "0.99"
-env_logger = "0.9"
-colored = "2.0.0"
-regex = "^1.5"
-serde_yaml = "0.8.23"
-hex = "0.4.2"
-serde = { version = "*", features = ["derive"] }
-serde_json = "*"
-log = "0.4.14"
-parity-scale-codec = { version = "*", features = ["derive"], default-features = false }
-sp-keyring = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-sp-core = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable" }
-sp-io = { version = "4.0.0-dev", git = "https://github.com/gear-tech/substrate.git", branch = "gear-stable", default-features = false }
-rayon = "1.5"
-
-# Internal deps
 gear-core = { path = "../core" }
-common = { package = "gear-common", path = "../common" }
-core-processor = { package = "gear-core-processor", path = "../core-processor", default-features = false }
-gear-backend-common = { path = "../core-backend/common" }
 gear-backend-wasmtime = { path = "../core-backend/wasmtime" }
+core-processor = { package = "gear-core-processor", path = "../core-processor" }
 
-[[bin]]
-name = "gtest"
-path = "src/main.rs"
-
-[lib]
-
-[features]
+codec = { package = "parity-scale-codec", version = "2", features = ["derive"] }
+hex = "0.4.3"
+colored = "2.0.0"
+env_logger = "0.9.0"
+path-clean = "0.1.0"
