+++ scripts/src/init.sh
@@ -33,14 +33,14 @@ wasm_init() {
 # $1 = ROOT_DIR
 js_init() {
   npm --prefix "$1"/utils/wasm-proc/metadata-js install
-  npm --prefix "$1"/gtest/src/js install
+  npm --prefix "$1"/gear-test/src/js install
   npm --prefix "$1"/rpc-tests install
 }
 
 # $1 = ROOT_DIR
 js_update() {
   npm --prefix "$1"/utils/wasm-proc/metadata-js update
-  npm --prefix "$1"/gtest/src/js update
+  npm --prefix "$1"/gear-test/src/js update
   npm --prefix "$1"/rpc-tests update
 }
